+++ azgo/aggr-get-iter.go
@@ -1,4 +1,4 @@
-// Copyright 2016 NetApp, Inc. All Rights Reserved.
+// Copyright 2017 NetApp, Inc. All Rights Reserved.
 
 package azgo
 
@@ -78,12 +78,19 @@ func (o *AggrGetIterRequest) ExecuteUsing(zr *ZapiRunner) (AggrGetIterResponse, 
 				o.SetTag(*nextTagPtr)
 			}
 
-			recordsRead := n.Result.NumRecords()
-			if recordsRead == 0 {
+			if n.Result.NumRecordsPtr == nil {
 				done = true
+			} else {
+				recordsRead := n.Result.NumRecords()
+				if recordsRead == 0 {
+					done = true
+				}
 			}
 
-			combined.Result.SetAttributesList(append(combined.Result.AttributesList(), n.Result.AttributesList()...))
+			if n.Result.AttributesListPtr != nil {
+				combined.Result.SetAttributesList(append(combined.Result.AttributesList(), n.Result.AttributesList()...))
+			}
+
 			if done == true {
 				combined.Result.ResultErrnoAttr = n.Result.ResultErrnoAttr
 				combined.Result.ResultReasonAttr = n.Result.ResultReasonAttr