+++ azgo/system-node-get-iter.go
@@ -1,4 +1,4 @@
-// Copyright 2016 NetApp, Inc. All Rights Reserved.
+// Copyright 2017 NetApp, Inc. All Rights Reserved.
 
 package azgo
 
@@ -87,7 +87,10 @@ func (o *SystemNodeGetIterRequest) ExecuteUsing(zr *ZapiRunner) (SystemNodeGetIt
 				}
 			}
 
-			combined.Result.SetAttributesList(append(combined.Result.AttributesList(), n.Result.AttributesList()...))
+			if n.Result.AttributesListPtr != nil {
+				combined.Result.SetAttributesList(append(combined.Result.AttributesList(), n.Result.AttributesList()...))
+			}
+
 			if done == true {
 				combined.Result.ResultErrnoAttr = n.Result.ResultErrnoAttr
 				combined.Result.ResultReasonAttr = n.Result.ResultReasonAttr